requirements.txt
streamlit
pandas
google-cloud-firestore
google-oauth2-service_account
json